refactor: Extract row & cell functions into helper
