+++ common/presenters/date-table/locations-to-population-table-component.js
@@ -1,63 +1,9 @@
-const { parseISO, isValid, isDate, addDays, format } = require('date-fns')
+const { parseISO, isValid, isDate, addDays } = require('date-fns')
 const { times } = require('lodash')
 
-const i18n = require('../../../config/i18n')
 const tablePresenters = require('../table')
 
-const freeSpaceCellContent = function ({ date, populationIndex = 0 }) {
-  return data => {
-    const { id: locationId } = data
-    const { free_spaces: count } =
-      data?.meta?.populations?.[populationIndex] || {}
-
-    const link =
-      count === undefined
-        ? i18n.t('population::add_space')
-        : i18n.t('population::spaces_with_count', { count })
-
-    const editQuicklink = count === undefined ? '/edit' : ''
-
-    const url = `/population/day/${format(
-      date,
-      'yyyy-MM-dd'
-    )}/${locationId}${editQuicklink}`
-
-    return `<a href="${url}">${link}</a>`
-  }
-}
-
-const dayConfig = function ({ date, populationIndex = 0 }) {
-  return {
-    head: {
-      date,
-      attributes: {
-        width: '80',
-      },
-      text: `Day ${populationIndex + 1}`,
-    },
-    row: {
-      date,
-      html: data => freeSpaceCellContent({ date, populationIndex })(data),
-    },
-  }
-}
-
-const establishmentConfig = {
-  head: {
-    html: 'population::labels.establishment',
-    attributes: {
-      width: '220',
-    },
-  },
-  row: {
-    attributes: {
-      scope: 'row',
-    },
-    html: data => {
-      return `<a>${data.title}</a>`
-    },
-  },
-}
+const { dayConfig, establishmentConfig } = require('./population-helpers')
 
 function locationsToPopulationTable({ query, startDate, dayCount = 5 } = {}) {
   let startDateAsDate = isDate(startDate) ? startDate : parseISO(startDate)
@@ -85,6 +31,4 @@ function locationsToPopulationTable({ query, startDate, dayCount = 5 } = {}) {
 
 module.exports = {
   locationsToPopulationTable,
-  freeSpaceCellContent,
-  dayConfig,
 }
